backend/requirements.txt
# FastAPI and Server
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-multipart>=0.0.9

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.10.1

# Configuration
python-dotenv>=1.0.1
pydantic-settings>=2.0.0

# Email validation
email-validator>=2.2.0

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code quality
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# HTTP requests
requests>=2.31.0

# Time zone handling
tzdata>=2024.2

# CLI tools
typer>=0.9.0

# Additional security
cryptography>=42.0.8

# Async testing
pytest-mock>=3.12.0
